db connection. Remove useless log msg
Former-commit-id: ae29f7f9948ae96ac0a3033744df28e36c3fa070

+++ MIB/src/edu/columbia/psl/cc/pojo/HotZone.java
@@ -14,7 +14,7 @@ public class HotZone {
 	
 	private InstNode endInst;
 	
-	private double levDist;
+	private int levDist;
 	
 	private double similarity;
 	
@@ -68,11 +68,11 @@ public class HotZone {
 		this.endInst = endInst;
 	}
 
-	public double getLevDist() {
+	public int getLevDist() {
 		return levDist;
 	}
 
-	public void setLevDist(double levDist) {
+	public void setLevDist(int levDist) {
 		this.levDist = levDist;
 	}
